A first stabl at a Jetpack Compose desktop app.
This is KLife, a Conway's Game of Life simulator, in a jetpack compose desktop window.

At present it lets you scale a field, resize the field, implements a toroidal world topology, standard Conway rules, and a start/stop button, plus a 'step' button for single manual stepping.

The window resizing is wierd - it breaks the field if you pull the mouse smaller than the resized window, while dragging the resize, and if you go far enough, starts throwing exceptions around negative padding numbers, etc. Something funky about resize and limits, etc.  IT sort of prevents too-small resizes, but it's hacky.

Also, you can edit the field by clicking on it, toggling the bit on and off, while the simulation is paused.

This is the first time I've used Compose, so it's probably terrible. I'm sure dispatching mutable state updates from a Timer is the wrong move, but it was a way to generate 'ticks' so I went for it.

It also can load a bitmap file, doing a terrible job of interpolating color into a black and white on-and-off image. The goal was to use pngs to store the bitmap, since the field is essentially that, but I may back that out, and just store it in a compressed text-file or a proto or something.